Reformat Code

+++ jsrc/Turtle.java
@@ -1,7 +1,7 @@
 class Turtle extends Herbivore {
 	private Race race;
 	
-	public Turtle( Universe universe, int x, int y, int currentAge ){
+	public Turtle( Universe universe, int x, int y, float currentAge ){
 		super(universe,x,y,currentAge);
 		universe.board.SetEl(ch(),y,x);
 		race = null;
